Implement DistDivScalar Rewrite Rule
(X / Broadcast(Y)) + (Z / Broadcast(Y)) => (X + Z) / Broadcast(Y),
	where Y is a scalar

TODO: This rule does not fire if Y is a known constant.
This is because in that case, the AS first rewrites [X / Broadcast(Y)] to
[X * (1 / Broadcast(Y)].
Interestingly, the AS is further expected to apply A*C+B*C => (A+B)*C, but does not.